requirements.txt
streamlit
python-dotenv
requests
konlpy
JPype1
aiohttp
google-genai
datasets
faiss-cpu
sentence-transformers
deepeval
langchain-community